+++ src/pyquickhelper/pycode/__init__.py
@@ -7,6 +7,7 @@ from .code_helper import remove_extra_spaces_and_pep8, remove_extra_spaces_folde
 from .coverage_helper import publish_coverage_on_codecov
 from .pip_helper import get_packages_list, get_package_info
 from .py3to2 import py3to2_convert_tree, py3to2_convert
+from .readme_helper import clean_readme
 from .setup_helper import process_standard_options_for_setup, write_version_for_setup, process_standard_options_for_setup_help
 from .setup_helper import available_commands_list
 from .tkinter_helper import fix_tkinter_issues_virtualenv